style: fix Prettier formatting
