+++ src/lib/heatmap-helpers.ts
@@ -1,0 +1,7 @@
+export function heatmapColor(count: number): string {
+  if (count === 0) return 'bg-slate-200';
+  if (count <= 2) return 'bg-blue-400';
+  if (count <= 5) return 'bg-blue-500';
+  if (count <= 10) return 'bg-indigo-600';
+  return 'bg-indigo-800';
+}
