+++ src/pages/SessionPage.tsx
@@ -1,5 +1,4 @@
 import { useEffect, useState, useCallback, useRef } from 'react';
-import { Link } from 'react-router-dom';
 
 import {
   getRandomCompetenciesAction,
@@ -160,7 +159,7 @@ export function SessionPage() {
         } else {
           setError(
             result.error ??
-            'Failed to load user information. Please try again later.'
+              'Failed to load user information. Please try again later.'
           );
         }
       } catch (err) {
@@ -201,7 +200,7 @@ export function SessionPage() {
         if (!result.success) {
           setError(
             result.error ??
-            'An unexpected error occurred while fetching the next task.'
+              'An unexpected error occurred while fetching the next task.'
           );
           if (isInitialLoad) setCompetencies([]);
           setIsTransitioning(false);
@@ -246,7 +245,7 @@ export function SessionPage() {
         if (!compResult.success || !compResult.competencies?.length) {
           setError(
             compResult.error ??
-            'Failed to fetch competency for resource mapping.'
+              'Failed to fetch competency for resource mapping.'
           );
           if (isInitialLoad) setCompetencies([]);
           setIsTransitioning(false);
@@ -256,7 +255,7 @@ export function SessionPage() {
         if (!resourceResult.success || !resourceResult.resource) {
           setError(
             resourceResult.error ??
-            'Failed to fetch learning resource. Make sure resources are seeded.'
+              'Failed to fetch learning resource. Make sure resources are seeded.'
           );
           setCompetencies(compResult.competencies);
           setLearningResource(null);
@@ -305,9 +304,9 @@ export function SessionPage() {
             const voteOpts =
               isSwapped && competencies && competencies.length >= 2
                 ? {
-                  originId: competencies[0]!.id,
-                  destinationId: competencies[1]!.id,
-                }
+                    originId: competencies[0]!.id,
+                    destinationId: competencies[1]!.id,
+                  }
                 : { relationshipId: currentRelationshipId! };
             const result = await submitCompetencyVoteAction(
               userId,
@@ -686,9 +685,10 @@ export function SessionPage() {
                   }}
                   className={`
                     h-full px-3 text-xs font-semibold rounded-md transition-all duration-200 flex items-center justify-center
-                    ${mappingMode === 'competency'
-                      ? 'bg-white text-slate-900 shadow-sm'
-                      : 'text-slate-500 hover:text-slate-700'
+                    ${
+                      mappingMode === 'competency'
+                        ? 'bg-white text-slate-900 shadow-sm'
+                        : 'text-slate-500 hover:text-slate-700'
                     }
                   `}
                 >
@@ -703,9 +703,10 @@ export function SessionPage() {
                   }}
                   className={`
                     h-full px-3 text-xs font-semibold rounded-md transition-all duration-200 flex items-center justify-center
-                    ${mappingMode === 'resource'
-                      ? 'bg-white text-slate-900 shadow-sm'
-                      : 'text-slate-500 hover:text-slate-700'
+                    ${
+                      mappingMode === 'resource'
+                        ? 'bg-white text-slate-900 shadow-sm'
+                        : 'text-slate-500 hover:text-slate-700'
                     }
                   `}
                 >
@@ -988,8 +989,8 @@ export function SessionPage() {
                             {relation === 'UNRELATED'
                               ? 'is unrelated to'
                               : RELATIONSHIP_TYPES.find(
-                                rt => rt.value === relation
-                              )?.label.toLowerCase() || ''}
+                                  rt => rt.value === relation
+                                )?.label.toLowerCase() || ''}
                           </span>{' '}
                           <span className="font-bold text-slate-900">
                             {competencies[1]!.title}
@@ -1165,9 +1166,10 @@ export function SessionPage() {
                       className={`
                         relative flex items-center justify-center gap-2 px-6 py-3 rounded-xl text-sm font-bold text-white
                         min-w-[180px]
-                        ${mappingMode === 'resource'
-                          ? 'bg-gradient-to-r from-pink-600 to-pink-500 shadow-lg shadow-pink-500/20 hover:shadow-xl hover:shadow-pink-500/30'
-                          : 'bg-gradient-to-r from-[#0a4da2] to-[#5538d1] shadow-lg shadow-blue-500/20 hover:shadow-xl hover:shadow-blue-500/30'
+                        ${
+                          mappingMode === 'resource'
+                            ? 'bg-gradient-to-r from-pink-600 to-pink-500 shadow-lg shadow-pink-500/20 hover:shadow-xl hover:shadow-pink-500/30'
+                            : 'bg-gradient-to-r from-[#0a4da2] to-[#5538d1] shadow-lg shadow-blue-500/20 hover:shadow-xl hover:shadow-blue-500/30'
                         }
                         transition-all duration-200 ease-out
                         hover:scale-[1.02]
